refactor: move Authorization class to root package
