+++ migration/identity-migration/src/main/java/io/camunda/migration/identity/config/AuthConfig.java
@@ -8,7 +8,7 @@
 package io.camunda.migration.identity.config;
 
 import io.camunda.security.auth.Authentication;
-import io.camunda.zeebe.auth.impl.Authorization;
+import io.camunda.zeebe.auth.Authorization;
 import java.util.Map;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
