+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/AuthorizationCheckBehaviorTest.java
@@ -7,9 +7,9 @@
  */
 package io.camunda.zeebe.engine.processing.authorization;
 
-import static io.camunda.zeebe.auth.impl.Authorization.AUTHORIZED_ANONYMOUS_USER;
-import static io.camunda.zeebe.auth.impl.Authorization.AUTHORIZED_USER_KEY;
-import static io.camunda.zeebe.auth.impl.Authorization.USER_TOKEN_CLAIM_PREFIX;
+import static io.camunda.zeebe.auth.Authorization.AUTHORIZED_ANONYMOUS_USER;
+import static io.camunda.zeebe.auth.Authorization.AUTHORIZED_USER_KEY;
+import static io.camunda.zeebe.auth.Authorization.USER_TOKEN_CLAIM_PREFIX;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
